IDE-1327 Fix broken terminate in waveforms other than the chalkboard; code de-duplication
The terminate command (in the waveform diagram) would only terminate components in the chalkboard waveform.
Terminating a component in some other waveform would attempt to terminate a component of the same name in
the chalkboard.

Extracted the terminate feature from the SAD and DCD diagrams to a single class and simplified the code.