Fix problem in test utility method that launches waveforms in the sandbox
It seems with SWTBot that nesting menu calls is problematic. When the second menu call is
invoked, the menu which was originally found gets disposed (or something similar), causing
an error. It's not the best solution, but you can go directly to the desired menu item, even
if it's a submenu.