Name components name_number when launching in Chalkboard
Previously, the first component launched in the Chalkboard got named "foo". The
second one would be "foo_1". This is inconsistent with design-time time behavior
for waveforms, where they would be named "foo_1", "foo_2". It's also inconsistent
with dragging from the palette onto the Chalkboard diagram. Lastly, it was making
tests a bit more confusing when testing the various way things are launched.