IDE-1332 Install waveforms/nodes to a location based on the name in their XML, not the project name
Eclipse gets the project name from the directory's name, which is problematic when someone has renamed
the directory that contains a project. We instead inspect the name element in the XML file to determine
the correct installation path.

Note that this commit also eliminates installation of the old GMF waveform diagram file.