IDE-1323 IDE-1324 Diagrams opened from the Target SDR might say they were for something in the workspace instead
The code for design-time diagrams was looking for the value of the SDRROOT environment variable in the URI
for the diagram to determine if it was opened from the workspace or the target SDR. That's wrong since the IDE
doesn't necessarily use the SDRROOT environment variable (it's configurable), and it had problems with paths
that made use of '..'. Instead, the code now checks the URI scheme, which is the correct way.