IDE-1330 IDE-1329 Issues with port coloring for external ports and port statistics
This is a complete overhaul of much of the port styling code. Much of the code now lives in PortStyleUtil.

The locking mechanisms in StyleUtil have been dropped. Now there's a property set on the Diagram that can be
checked by the utility method DUtil.isConnecting() to know if there's a conenction in progress.

In the long term, some kind of a decorator on the port to change its color for connections-in-progress and
for port statistics might be best. By changing the color on the ports with a model command, we're disabling
undo functionality. The problem where the color for a port statistic get cleared is also still present.