IDE-1380 Correct a race condition with SDR root / IDL library model loading
The code which put the SDR root model object and the IDL library model object inside a Resource,
inside an EditingDomain was deferred to being executed in an Eclipse job. This meant that even
after the bundle providing the model was started, the model objects weren't in the EditingDomain
for a brief span of time. This could cause issues with any code that accessed the SDR root or the
IDL library early on during startup.