IDE-1366 Refresh Graphiti diagram in a read-only transaction to prevent other threads from changing the model
The Graphiti command stack implementation contains a bug wherein a command that should otherwise create a new read/write transaction is allowed to run even though the currently active transaction is owned by another thread and/or is read-only. We work around this by overriding execute() in our own command stack implementation and explicitly creating a transaction in the problematic case.