Improvements and fixes for sandbox resource termination
Backport parts of 7a37252 and 76adb21.

These fixes ensure the IDE schedules termination jobs when model objects of local
resources are disposed. Previously many of the dispose() methods would invoke releaseObject().

The IDE also allows the termination jobs a little bit of time to complete if the IDE is
shutting down.