IDE-1452 Correct DCD diagram being dirty immediately after open
Previous to the work for IDE-1452, the initial creation of the diagram input happened in the context
of a NonDirtyingCommand (a command that is not undoable, and does not dirty the diagram). The first
commit in the IDE-1452 series removed this. This commit changes a command in the DCD editor's code
to be a NonDirtyingCommand, which matches the SAD implementation, and prevents the editor from being
dirty immediately after it is opened.