IDE-1189 Restructure custom features, hover pad, context menus for Graphiti diagrams
This change adds the "group.control" group/separator into the graphiti context
menus, allowing extension points to contribute to that area. Currently things
like "Connect", "Start", "Stop", etc all fall in this group within the explorer
view. Thus, the existing menus and handlers are now automatically resused by
Graphiti.

Accordingly this change drops many of the custom features that duplicated
command handlers used by the explorer view. The only exception to this are
custom features that are used by the hover pad, which is a unique issue that
Graphiti must handle.