IDE-1911 Correct launch error for services
The previous fix that refuses a configure() call if it is inappropriate for a
service began causing a launch error for services that did not support
configure(), because ALL services were getting configure() called on them, not
just the ones that support it. Now we'll only call it on the ones that support
it.

Also, the configure() call is now relocated to the appropriate place, within
the registration call to the device manager.