IDE-2170 Correct editing of external prop IDs in SAD editor
Previously the cell editor for a property's external ID was a combobox. The
combobox had two entries: "" (i.e. unset) and the property's name (which isn't
necessarily a unique external property name). Further, the text on the combo
was set to a unique name, which doesn't necessarily match the two choices. This
is now much simpler - we use a text box and populate it with a suggested unique
name.

The validation of the cell editor's text now checks the assembly controller's
properties for a collision.

An extreme edge case in construction of the unique external property ID was
corrected.