IDE-2198 Better file choice dialogs
The previous dialog for choosing an SPD/SAD XML file from the workspace was
confusing because it allowed seleceting multiple files, showed all files, and
had a split pane between folders and files. The new one is greatly simplified.
Just a tree of the workspace root. Files are filtered so on the appropriate
kind (SPD/SAD XML) are shown, and only one can be selected.

Also, the dialog for choosing a file from the file system now opens to an
appropriate location in $SDRROOT, which should help most people get to the
directory they actually want more quickly.