Adjust test cleanup for dialogs left open
Our cleanup code was forcefully closing dialogs left open by a previous test by
calling Shell.dispose(). This appeared to cause problems for the dialog that
were closed in this way. Specifically, this was seen with the run configuration
tests just introduced for IDE-929, IDE-2197 and IDE-2198. After a
Shell.dispose() on the run configurations dialog, Eclipse wouldn't open the
dialog any more. It was like the menus had no effect.

It's possible this has had negative effects on other areas of the IDE as well
in our testing. The code is being removed in favor of just using SWTBot's
cleanup code. This code originally had a bug in it, which may be why we had to
write our own code, but this has been corrected since SWTBot 2.6.0.