IDE-1107 Ensure local comps added to corresponding domain waveform
The trigger points for the ScaDebugContentProvider navigator content needed to
include LocalScaComponent. Without this, the sequence goes like this:
* Viewer tries to find children of ScaWaveform
* LocalComponentContentProvider content provider declares ScaWaveform in its
  trigger points
* LocalComponentContentProvider contributes LocalScaComponent(s) to the
  ScaWaveform
* The user expands the LocalScaComponent to see its children
* Viewer tries to find children of LocalScaComponent
* Viewer finds NO matching trigger points (will NOT ask ScaDebugContentProvider
  for children, which is what would provide them)
* Viewer asks the original content provider that gave it the LocalScaComponent
  (i.e. LocalComponentContentProvider) for any children, but it doesn't supply
  any.

Debugging to see this occur can be confusing if you've browsed around the
sandbox a little. The common navigator remembers what content providers yield
children for different types. This means that if you expand a LocalScaComponent
in the sandbox, the viewer will remember that ScaDebugContentProvider can
provide children for a LocalScaComponent. The next time you expand a
LocalScaComponent that is a child of an ScaWaveform, it will ask both
ScaDebugContentProvider and LocalComponentContentProvider for children.

Providing the trigger point ensures the navigator knows about the children that
can be provided right off the bat.