IDE-2222 Don't call unload after changing SDR root info
Previously the code would unload() to unload and remove all the contents of the
SDR root model. This made some sense since it would prevent the SDR model from
being inconsistent (i.e. the location is set to /bar, but all the files in the
model were loaded from /foo).

However, unload() is synchronized AND it runs EMF commands. This meant that the
order the two locks were being used wasn't consistent and lead to deadlock.
Also, while calling unload() might have made the model self-consistent, it
didn't change the fact that the model was probably loaded with content before,
and now just has no content because the content was reloaded.

By dropping the call to unload(), we can simply let the caller who is changing
the SDR root location then call reload(), which was already occuring anyway in
an Eclipse Job.