IDE-2237 Minimize code while processing notification
SdrResourceFactoryProvider had a lot of code called within the scope of the
notification that SPDs were added or removed. Notifications are sent within the
scope of a transaction, which meant a lock was being held for longer than
necessary.

SdrResourceFactoryProvider now places notifications on a queue for processing
by an Eclipse Job.