IDE-2271 Ensure componentinstantiation id from SAD is used when launching component
In 2.1.0, a code change made the IDE stop passing the component instantiation
ID from the SAD file in one of the code paths used to launch components. This
meant the ID in the SAD file wasn't used to create the component's identifier
(i.e. CF.Resource#identifier()), and therefore when the SCA model loads the
component's instantiation identifier can't be computed. This prevents making
connections specified in the SAD file when the waveform is launched in the
sandbox.